Rearranged image folder & code
Images were being called and displayed correctly within the IDE but not
when exported into a .jar file or .exe. Rearranged the img folder to
within the source folder, implemented imageURL with
getClass.getResource() methods to call the file path correctly within
the jar. Had to rewrite certain areas to fix the issue